feat: remove heading tooltip in RuleDetails page
