+++ web/src/pages/RuleDetails/RuleDetails.test.tsx
@@ -29,6 +29,7 @@ import {
   waitFor,
   waitForElementToBeRemoved,
   waitMs,
+  within,
 } from 'test-utils';
 import { DEFAULT_LARGE_PAGE_SIZE, DEFAULT_SMALL_PAGE_SIZE } from 'Source/constants';
 import {
@@ -272,7 +273,7 @@ describe('RuleDetails', () => {
       }),
     ];
 
-    const { getByText, getByTestId, getByAriaLabel, getAllByText } = render(
+    const { getByText, getByTestId } = render(
       <Route exact path={urls.logAnalysis.rules.details(':id')}>
         <RuleDetails />
       </Route>,
@@ -288,15 +289,14 @@ describe('RuleDetails', () => {
     fireEvent.click(getByText('Rule Matches'));
 
     const loadingListingInterfaceElement = getByTestId('rule-alerts-listing-loading');
-    expect(loadingListingInterfaceElement).toBeTruthy();
     await waitForElementToBeRemoved(loadingListingInterfaceElement);
-    expect(getByText('Alert 1')).toBeInTheDocument();
-    expect(getByText('Rule Match')).toBeInTheDocument();
+    const withinTabPanel = within(getByTestId('rule-matches-tabpanel'));
+    expect(withinTabPanel.getByText('Alert 1')).toBeInTheDocument();
+    expect(withinTabPanel.getByText('Rule Match')).toBeInTheDocument();
 
-    expect(getAllByText('Destinations').length).toEqual(2);
-    expect(getByText('Log Types')).toBeInTheDocument();
-    expect(getByText('Events')).toBeInTheDocument();
-    expect(getByAriaLabel('Change Alert Status')).toBeInTheDocument();
+    expect(withinTabPanel.getByText('Destinations')).toBeInTheDocument();
+    expect(withinTabPanel.getByText('Log Types')).toBeInTheDocument();
+    expect(withinTabPanel.getByText('Events')).toBeInTheDocument();
   });
 
   it('fetches the alerts matching the rule errors', async () => {
@@ -341,7 +341,7 @@ describe('RuleDetails', () => {
       }),
     ];
 
-    const { getByText, getByTestId, getByAriaLabel, getAllByText } = render(
+    const { getByText, getByTestId } = render(
       <Route exact path={urls.logAnalysis.rules.details(':id')}>
         <RuleDetails />
       </Route>,
@@ -357,15 +357,13 @@ describe('RuleDetails', () => {
     fireEvent.click(getByText('Rule Errors'));
 
     const loadingListingInterfaceElement = getByTestId('rule-alerts-listing-loading');
-    expect(loadingListingInterfaceElement).toBeTruthy();
     await waitForElementToBeRemoved(loadingListingInterfaceElement);
-    expect(getByText('Error 1')).toBeInTheDocument();
-    expect(getByText('Rule Error')).toBeInTheDocument();
-
-    expect(getAllByText('Destinations').length).toEqual(2);
-    expect(getByText('Log Types')).toBeInTheDocument();
-    expect(getByText('Events')).toBeInTheDocument();
-    expect(getByAriaLabel('Change Alert Status')).toBeInTheDocument();
+    const withinTabPanel = within(getByTestId('rule-errors-tabpanel'));
+    expect(withinTabPanel.getByText('Error 1')).toBeInTheDocument();
+    expect(withinTabPanel.getByText('Rule Error')).toBeInTheDocument();
+    expect(withinTabPanel.getByText('Destinations')).toBeInTheDocument();
+    expect(withinTabPanel.getByText('Log Types')).toBeInTheDocument();
+    expect(withinTabPanel.getByText('Events')).toBeInTheDocument();
   });
 
   it('fetches the alerts matching the rule & shows an empty fallback if no alerts exist', async () => {
