+++ web/src/pages/RuleDetails/RuleDetailsBanner/index.tsx
@@ -16,4 +16,4 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-export { default } from './RuleCardDetails';
+export { default } from './RuleDetailsBanner';
